fix(desktop): PR Changes diff 接入语法高亮
抽取 tokenizeDiffHunks 供工具卡与 PR diff 共用；ReviewCommentHunkView
经 monacoLanguageId 解析语言并传入 refractor tokens。
